Move JSONRawObjectWriter to pkg/util/jsonrawobjectwriter
