+++ pkg/util/jsonrawobjectwriter/json_raw_object_writer_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019 Datadog, Inc.
 
-package jsonstream
+package jsonrawobjectwriter
 
 import (
 	"testing"
